+++ mfs/file.go
@@ -3,7 +3,9 @@ package mfs
 import (
 	"context"
 	"errors"
+	"os"
 	"sync"
+	"time"
 
 	dag "github.com/ipfs/boxo/ipld/merkledag"
 	ft "github.com/ipfs/boxo/ipld/unixfs"
@@ -176,4 +178,102 @@ func (fi *File) Sync() error {
 // Type returns the type FSNode this is
 func (fi *File) Type() NodeType {
 	return TFile
+}
+
+func (fi *File) Mode() (os.FileMode, error) {
+	fi.nodeLock.RLock()
+	defer fi.nodeLock.RUnlock()
+
+	nd, err := fi.GetNode()
+	if err != nil {
+		return 0, err
+	}
+	fsn, err := ft.ExtractFSNode(nd)
+	if err != nil {
+		return 0, err
+	}
+	return fsn.Mode() & 0xFFF, nil
+}
+
+func (fi *File) SetMode(mode os.FileMode) error {
+	nd, err := fi.GetNode()
+	if err != nil {
+		return err
+	}
+
+	fsn, err := ft.ExtractFSNode(nd)
+	if err != nil {
+		if errors.Is(err, ft.ErrNotProtoNode) {
+			// Wrap raw node in protonode.
+			data := nd.RawData()
+			return fi.setNodeData(ft.FilePBDataWithStat(data, uint64(len(data)), mode, time.Time{}))
+		}
+		return err
+	}
+
+	fsn.SetMode(mode)
+	data, err := fsn.GetBytes()
+	if err != nil {
+		return err
+	}
+
+	return fi.setNodeData(data)
+}
+
+// ModTime returns the files' last modification time
+func (fi *File) ModTime() (time.Time, error) {
+	fi.nodeLock.RLock()
+	defer fi.nodeLock.RUnlock()
+
+	nd, err := fi.GetNode()
+	if err != nil {
+		return time.Time{}, err
+	}
+	fsn, err := ft.ExtractFSNode(nd)
+	if err != nil {
+		return time.Time{}, err
+	}
+	return fsn.ModTime(), nil
+}
+
+// SetModTime sets the files' last modification time
+func (fi *File) SetModTime(ts time.Time) error {
+	nd, err := fi.GetNode()
+	if err != nil {
+		return err
+	}
+
+	fsn, err := ft.ExtractFSNode(nd)
+	if err != nil {
+		if errors.Is(err, ft.ErrNotProtoNode) {
+			// Wrap raw node in protonode.
+			data := nd.RawData()
+			return fi.setNodeData(ft.FilePBDataWithStat(data, uint64(len(data)), 0, ts))
+		}
+		return err
+	}
+
+	fsn.SetModTime(ts)
+	data, err := fsn.GetBytes()
+	if err != nil {
+		return err
+	}
+
+	return fi.setNodeData(data)
+}
+
+func (fi *File) setNodeData(data []byte) error {
+	nd := dag.NodeWithData(data)
+	err := fi.inode.dagService.Add(context.TODO(), nd)
+	if err != nil {
+		return err
+	}
+
+	fi.nodeLock.Lock()
+	defer fi.nodeLock.Unlock()
+	fi.node = nd
+	parent := fi.inode.parent
+	name := fi.inode.name
+
+	return parent.updateChildEntry(child{name, fi.node})
 }